refactor(run-query): makes code more testable and separates code
refactor improves logic and flow
