+++ test/testSend.js
@@ -12,18 +12,20 @@
 
 /* eslint-env mocha */
 const assert = require('assert');
+const sinon = require('sinon');
 const { AssertionError } = require('assert');
-const util = require('../src/util.js');
-const send = require('../src/sendquery.js');
+const env = require('../src/env.js');
+const utils = require('../src/util.js');
+const { execute } = require('../src/sendquery.js');
 
 describe('bigquery tests', () => {
   it('runs a query', async () => {
-    const { results } = await send.execute(util.email, util.key, util.projectid, '/list-everything', '0bxMEaYAJV6SoqFlbZ2n1f');
+    const { results } = await execute(env.email, env.key, env.projectid, 'list-everything', '0bxMEaYAJV6SoqFlbZ2n1f');
     assert.ok(Array.isArray(results));
   }).timeout(5000);
 
   it('runs a query with params', async () => {
-    const { results } = await send.execute(util.email, util.key, util.projectid, 'list-everything', '0bxMEaYAJV6SoqFlbZ2n1f', {
+    const { results } = await execute(env.email, env.key, env.projectid, 'list-everything', '0bxMEaYAJV6SoqFlbZ2n1f', {
       limit: 10,
     });
     assert.ok(Array.isArray(results));
@@ -32,7 +34,7 @@ describe('bigquery tests', () => {
 
   it('throws without projectid', async () => {
     try {
-      await send.execute(util.email, util.key, undefined, 'list-everything', '0bxMEaYAJV6SoqFlbZ2n1f');
+      await execute(env.email, env.key, undefined, 'list-everything', '0bxMEaYAJV6SoqFlbZ2n1f');
       assert.fail('expected exception not thrown');
     } catch (e) {
       if (e instanceof AssertionError) {
@@ -43,10 +45,13 @@ describe('bigquery tests', () => {
   }).timeout(5000);
 
   it('throws with bad query', async () => {
+    const fakeQuery = '# this query is intentionally broken.';
+    sinon.stub(utils, 'loadQuery').returns(fakeQuery);
     try {
-      await send.execute(util.email, util.key, util.projectid, 'break-something', '0bxMEaYAJV6SoqFlbZ2n1f');
+      await execute(env.email, env.key, env.projectid, 'break-something', '0bxMEaYAJV6SoqFlbZ2n1f');
       assert.fail('expected exception not thrown');
     } catch (e) {
+      sinon.restore();
       if (e instanceof AssertionError) {
         // bubble up the assertion error
         throw e;
@@ -57,15 +62,14 @@ describe('bigquery tests', () => {
 
 describe('sql loading and processing', () => {
   it('loadQuery loads a query', () => {
-    const result = send.loadQuery('list-everything');
+    const result = utils.loadQuery('list-everything');
     assert.ok(result.match(/select/i));
   });
 
   it('query parameters are processed', () => {
+    const fakeQuery = '--- helix-param: helix\n--- helix-param2: helix2\n--- helix-param3: helix3\n# this query is intentionally broken.';
     const EXPECTED = { 'helix-param': 'helix', 'helix-param2': 'helix2', 'helix-param3': 'helix3' };
-    const query = send.loadQuery('break-something', './test');
-    const ACTUAL = send.getExtraParameters(query);
-
+    const ACTUAL = utils.getExtraParameters(fakeQuery);
     assert.deepEqual(EXPECTED, ACTUAL);
   });
 });
